Request class refactor
