+++ src/Request/CallNumberTrait.php
@@ -1,0 +1,21 @@
+<?php
+
+
+namespace Nexy\PayboxDirect\Request;
+
+
+trait CallNumberTrait
+{
+    /**
+     * @var int
+     */
+    private $callNumber;
+
+    /**
+     * @return int
+     */
+    final protected function getCallNumber()
+    {
+        return $this->callNumber;
+    }
+}
